Move XCM runtime configurations into their own files
